+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.22] - 2025-01-13
+
 ### Fix
 
 - Preserve comments when update payments
@@ -176,7 +178,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - ticketType to the list
 
-[unreleased]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/0.0.21...HEAD
+[unreleased]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/0.0.22...HEAD
+[0.0.22]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/v0.0.21...v0.0.22
 [0.0.21]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/v0.0.20...v0.0.21
 [0.0.20]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/v0.0.19...v0.0.20
 [0.0.19]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/v0.0.18...v0.0.19